challenges/backend/src/app/AuctionMonitorApp.ts
import { inject, injectable } from "inversify";
import { ILogger } from "./services/Logger/interface/ILogger";
import { DependencyIdentifier } from "./DependencyIdentifiers";
import "reflect-metadata";
import { ICarOnSaleClient } from "./services/CarOnSaleClient/interface/ICarOnSaleClient";

@injectable()
export class AuctionMonitorApp {
    public constructor(
        @inject(DependencyIdentifier.LOGGER) private logger: ILogger,
        @inject(DependencyIdentifier.CAR_ON_SALE) private readonly carOnSaleClient: ICarOnSaleClient
    ) {
    }


    public async start(): Promise<void> {
        this.logger.log(`Auction Monitor started.`);
        this.carOnSaleClient.print();

    }

}
